refactor: ACE가 11을 반환할 수 있다는 내용을 Symbol로 이동 및 점수계산 테스트 알맞은 클래스로 이동
1. getBiggerAceScore()메소드를 통해 의도를 드러냄
2. 점수계산 테스트 HandsTest -> CardsTest
